Move model definition to each route
